refactor: changes project architecture
The source code is now inside a module called "src" and the assets are
inside a folder called assets. The entry point of the game still the
main.py file in the root folder of the project, but it's only a "wrapper" for the main function of the views.py layer (old main.py).
With this new organization, it's easier to navigate the project and
understand it.